Extract UserPiiResolver, slim EventFactory (rename ServerEventFactory)
Split the old ServerEventFactory into a focused event builder plus a request
identity resolver, and give the builder a minimal public surface.

- EventFactory (renamed from ServerEventFactory): public surface is now just
  create( $event_name, $data = array(), $tracking_name = '', $prefer_referrer
  = false ) and create_from_array( $event ). safe_create_event is removed —
  callers run their own callback and pass the resulting array to create().
  new_event, split_user_data_and_custom_data, map_user_data_keys and
  get_request_uri are now private. create() only stamps
  fb_integration_tracking when a tracking name is given, so a bare
  create( $event_name ) yields a clean base event.
- UserPiiResolver (new): the single source of the current request's user
  identifiers — get_ip_address(), get_http_user_agent(), get_fbp(), get_fbc()
  (the old get_fbp_value/get_fbc_value wrappers collapse into these) plus the
  private fbclid/IP helpers. EventFactory pulls PII from here.
- FacebookPluginUtils::split_name(): moved here (not event-building).

Callers updated to the new API: Signals::track -> EventFactory::create;
FacebookCapiEvent and OpenBridge run their callback inline then create();
ServerEventAsyncTask -> create_from_array (dead convert_user_data removed);
ReleaseSignalsAjax -> UserPiiResolver::get_fbp/get_fbc and builds its base
event via EventFactory::create( $event_name ); MailChimp -> create() with
self::readFormData(). The 8 disabled integrations keep their stale
ServerEventFactory references (never loaded; converted in a later stage).

Tests: ServerEventFactoryTest split into UserPiiResolverTest (IP/UA/fbp/fbc)
and EventFactoryTest (event building via create()/create_from_array),
including coverage that create() wires the resolved IP/UA/fbp/fbc onto the
event and that create_from_array maps the short AAM user-data keys.
FacebookServerSideEventTest/OpenBridgeTest build fixtures via
EventFactory::create( $name ).

Full suite: 192 green; phpcs whole-tree clean.
