+++ core/class-releasesignalsajax.php
@@ -189,8 +189,8 @@ class ReleaseSignalsAjax {
 
         wp_send_json_success(
             array(
-                'fbp'        => ServerEventFactory::get_fbp_value(),
-                'fbc'        => ServerEventFactory::get_fbc_value(),
+                'fbp'        => UserPiiResolver::get_fbp(),
+                'fbc'        => UserPiiResolver::get_fbc(),
                 'sent_count' => count( $events_to_send ),
                 'user_info'  => $user_info,
             )
@@ -235,7 +235,7 @@ class ReleaseSignalsAjax {
      */
     private function build_event( $event_data ) {
         $event_name = sanitize_text_field( $event_data['event_name'] );
-        $event      = ServerEventFactory::new_event( $event_name );
+        $event      = EventFactory::create( $event_name );
 
         if ( ! empty( $event_data['event_time'] ) ) {
             $event->setEventTime( absint( $event_data['event_time'] ) );
